requirements.txt
browser-use[memory]==0.2.5
langchain-google-genai==2.1.2
python-dotenv==1.1.0
playwright==1.52.0
pydantic>=2.0.0,<3.0.0
httpx>=0.27.2
langchain-core==0.3.49
numpy>=1.24.0